- [x] improved build scripts
- [x] improved CMake installation scripts
- [x] added self-contained examples directory to demostrate the use of
UTAP
- [x] adjusted README.md accordingly
- [x] split GH CI scripts and added test for examples
- [x] moved compiler warnings and sanitizer options from cmake to GH CI
scripts so that examples would use exactly the same options.
- [x] fixed some assertions triggered by the example in Debug builds
- [x] added GH CI script AppleClang/Darwin and fixed some errors and
warnings

+++ include/utap/document.hpp
@@ -366,7 +366,7 @@ struct Template : public Instance, Declarations
     bool is_TA{true};
     bool is_instantiated{false};  ///< Is the template used in the system
 
-    int add_dynamic_eval(Expression t)
+    size_t add_dynamic_eval(Expression t)
     {
         dynamic_evals.push_back(std::move(t));
         return dynamic_evals.size() - 1;
